+++ include/vctr/vctr.h
@@ -151,19 +151,19 @@
 #include "Expressions/Filter/SIMDFilter.h"
 #include "Expressions/Filter/PlatformVectorOpsFilter.h"
 
-#include "Expressions/Core/Abs.h"
-#include "Expressions/Core/Clamp.h"
-#include "Expressions/Core/Cube.h"
-#include "Expressions/Core/Sqrt.h"
-#include "Expressions/Core/Square.h"
-#include "Expressions/Core/Add.h"
-#include "Expressions/Core/Subtract.h"
-#include "Expressions/Core/Multiply.h"
-#include "Expressions/Core/Divide.h"
-#include "Expressions/Core/Max.h"
-#include "Expressions/Core/Min.h"
-#include "Expressions/Core/Mean.h"
-#include "Expressions/Core/Sum.h"
+#include "Expressions/BasicMath/Abs.h"
+#include "Expressions/BasicMath/Clamp.h"
+#include "Expressions/BasicMath/Cube.h"
+#include "Expressions/BasicMath/Sqrt.h"
+#include "Expressions/BasicMath/Square.h"
+#include "Expressions/BasicMath/Add.h"
+#include "Expressions/BasicMath/Subtract.h"
+#include "Expressions/BasicMath/Multiply.h"
+#include "Expressions/BasicMath/Divide.h"
+#include "Expressions/BasicMath/Max.h"
+#include "Expressions/BasicMath/Min.h"
+#include "Expressions/BasicMath/Mean.h"
+#include "Expressions/BasicMath/Sum.h"
 
 #include "Expressions/Complex/Angle.h"
 #include "Expressions/Complex/Conjugate.h"
